backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
pydantic-settings>=2.0.0
openai>=1.0.0
azure-cognitiveservices-speech>=1.30.0
aiohttp>=3.8.0
python-multipart>=0.0.5
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # Required for TestClient
flake8>=6.0.0
pylint>=3.0.0
mypy>=1.0.0
pytest-cov>=4.0.0
bandit>=1.7.0
